Move button to component and align right
